fix: resolved import-path type due to rename-refactoring
